refactor event sources into their own modules
